backend/requirements.txt
# Image/Video converter backend - Python 3.13
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.17
Pillow==11.0.0
aiofiles==24.1.0
python-dotenv==1.0.1
sqlalchemy==2.0.36
pymysql==1.1.1
cryptography>=42.0.0